Resolve C4365 warnings (signed/unsigned mismatches)

+++ src/charls_jpegls_encoder.cpp
@@ -103,7 +103,7 @@ struct charls_jpegls_encoder final
     size_t estimated_destination_size() const
     {
         check_operation(is_frame_info_configured());
-        return checked_mul(checked_mul(checked_mul(frame_info_.width, frame_info_.height), frame_info_.component_count),
+        return checked_mul(checked_mul(checked_mul(frame_info_.width, frame_info_.height), static_cast<size_t>(frame_info_.component_count)),
                            bit_to_byte_count(frame_info_.bits_per_sample)) +
                1024 + spiff_header_size_in_bytes;
     }
